isolate test_recover_digest_mismatch
